Added color sort
The function takes the data from textarea, strips the '#' prefix, and creates a data array. Duplicate values are then removed and the data array is sorted in alphabetical order. 

Next, we loop through the array to create a list and prefix each value with a '#' and set the value of the textarea. (This should probably be an array map instead)

Next, we loop through the array to create the box previews using the div template and style them by calling setColors(). In retrospect, this function may not be necessary and should probably be removed in a future update.